requirements.txt
Django
djangorestframework
markdown
django-filter
mysql-python
djangorestframework-jwt
